requirements.txt
aniso8601==7.0.0
asgiref==3.7.2
Django==4.2.7
django-ckeditor==6.7.0
django-js-asset==2.1.0
Pillow==10.1.0
promise==2.3
Rx==1.6.3
singledispatch==4.1.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2023.3
